% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIS-package.R
\docType{package}
\name{BIS-package}
\alias{BIS}
\alias{BIS-package}
\title{BIS: PKG_TITLE}
\description{
PKG_DESC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/serkor1/bis-main}
  \item Report bugs at \url{https://github.com/serkor1/bis-main/issues}
}

}
\author{
\strong{Maintainer}: Serkan Korkmaz \email{serkor1@duck.com}

}
\keyword{internal}
