% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{clickable_card}
\alias{clickable_card}
\title{clickable_card}
\usage{
clickable_card(inputid, outputval, ...)
}
\arguments{
\item{inputid}{[character] of [length] 1. The id of the element. Can be wrapped in [NS].}

\item{outputval}{[character] of [length] 1. The outputted value when the card is clicked.}

\item{...}{parameters passed into [bslib::card()]-funciton.}
}
\value{
The return value, if any, from executing the utility.
}
\description{
Description
}
