% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_tabs}
\alias{create_tabs}
\title{Dynamically create multiple tabs}
\usage{
create_tabs(X, fn, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{fn}{A [function] to be passed into the body
of the panel}

\item{...}{}
}
\value{
A [shiny.tag]. More specifically a dynamically rendered
navigation panel.
}
\description{
Dynamically create multiple tabs
}
\examples{
# script: scr_renderTabs
# date: 2024-03-29
# author: Serkan Korkmaz, serkor1@duck.com
# objective: Demonstrate the use
# of tabs-function
# script start;

# Create tabs with
# some random content
create_tabs(
  X = 1:10,
  fn = function(x) {

    p(
      paste(
        "content of tab:", x
      )
    )

  },
  bslib::nav_item(
    actionButton(
      inputId = "create_tabs",
      label = "Button"
    )
  )
)


# script end;
}
