% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from SQL database}
\usage{
extract_data(DB_connection = NULL, table = "model1", ...)
}
\arguments{
\item{...}{}
}
\description{
Extract data from SQL database
}
