% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_bis}
\alias{get_bis}
\title{Download, parse, and read into memory a BIS data set}
\usage{
get_bis(url, ...)
}
\arguments{
\item{url}{The url to the data set to be read (usually acquired through get_dataset())}

\item{...}{Arguments passed to download.file (e.g. quiet = TRUE)}
}
\value{
A data frame
}
\description{
Download, parse, and read into memory a BIS data set
}
\examples{
\dontrun{
datasets <- get_datasets()
rates <- get_bis(datasets$url[16])
}
}
