% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_datasets}
\alias{get_datasets}
\title{Download data frame of available BIS data sets}
\usage{
get_datasets()
}
\value{
A data frame
}
\description{
Download data frame of available BIS data sets
}
\examples{
datasets <- get_datasets()
}
