% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{input_parameters}
\alias{input_parameters}
\title{Create a list of input values}
\usage{
input_parameters(DT, value = "value", label = "label", as_list = NULL)
}
\arguments{
\item{DT}{A [data.frame]-object containing the input parameters and
labels.}

\item{value}{A [character]-vector of length 1. The values of the variables}

\item{label}{A [character]-vector of length 1. The label of the variables}

\item{as_list}{A [logical]-vector of length 1. If [TRUE] the values are
split by label}
}
\description{
Create a list of input values
}
