% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{picker_input}
\alias{picker_input}
\title{Add a pickerinput}
\usage{
picker_input(
  inputid,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  search = TRUE,
  size = "auto"
)
}
\arguments{
\item{inputid}{A [character]-vector of [length] 1.}

\item{label}{A [character]-vector of [length] 1.}

\item{choices}{A [character]-vector of [length] N.}

\item{selected}{A [character]-vector of [length] N.}

\item{multiple}{A [logical]-vector of [length] 1.}

\item{search}{A [logical]-vector of [length] 1.}

\item{size}{A [numeric]-vector of [length] 1.}
}
\description{
A wrapper of the [shinyWidgets::pickerInput()]-function, with predefined
danish wording.
}
\details{
See [shinyWidgets::pickerInput()] for usage.
}
