% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{clickable_card}
\alias{clickable_card}
\title{Create an elevated and clickable card}
\usage{
clickable_card(inputid, outputval, ...)
}
\arguments{
\item{inputid}{[character] of [length] 1. The id of the element. Can be wrapped in [NS].}

\item{outputval}{[character] of [length] 1. The outputted value when the card is clicked.}

\item{...}{parameters passed into [bslib::card()]-function.}
}
\value{
A clickable [bslib::card()] with [class] `clickable-card`
}
\description{
An elevated card that is clickable.
}
