% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIS.R
\name{get_bis}
\alias{get_bis}
\title{Download and parse a BIS data set}
\usage{
get_bis(item_url, ...)
}
\arguments{
\item{item_url}{Character. URL of the data set to be imported (usually
obtained via \code{get_datasets()}).}

\item{...}{Arguments passed to \code{download.file()} (e.g.
\code{quiet = TRUE}).}
}
\value{
A tibble data frame.
}
\description{
Download and parse a BIS data set
}
\examples{
\donttest{
ds <- get_datasets()
df <- get_bis(ds$url[ds$id == "WS_CBPOL_csv_flat"])
}
}
