% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIS.R
\name{get_bis}
\alias{get_bis}
\title{Download and parse a BIS data set}
\usage{
get_bis(url, auto_pivot = TRUE, ...)
}
\arguments{
\item{url}{Character. URL of the data set to be imported (usually obtained
through \code{get_datasets()}).}

\item{auto_pivot}{Logical. Controls whether source data set is converted to
long format. Set this to \code{FALSE} to disable conversion (default: TRUE).}

\item{...}{Arguments passed to \code{download.file()} (e.g.
\code{quiet = TRUE}).}
}
\value{
A tibble data frame, or a list of tibble data frames in cases where
the source zip file contains multiple csv files.
}
\description{
Download and parse a BIS data set
}
\details{
Parsing may fail if the amount of available memory is insufficient
for executing the necessary pivot operation. This is the case with large data
sets in particular. As a workaround, users may wish to set
\code{auto_pivot = FALSE} when calling \code{get_bis()}, then subset the
data and run \code{pivot_longer_bis()} manually. See example 2 below.
}
\examples{
\dontrun{
# Example 1
ds <- get_datasets()
df <- get_bis(ds$url[2])

# Example 2
options(timeout = 600)
lbs <- get_bis(ds$url[(ds$id == "full_lbs_d_pub_csv")], auto_pivot = FALSE)
lbs <- subset(lbs, l_parent_cty \%in\% c("US", "DE", "JP"))
lbs <- pivot_longer_bis(lbs)
}
}
