% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIS.R
\name{get_datasets}
\alias{get_datasets}
\title{Retrieve a list of available BIS data sets}
\usage{
get_datasets(base_url = "https://data.bis.org/bulkdownload")
}
\arguments{
\item{base_url}{Character. URL of the BIS's homepage listing single file data
sets for download (optional).}
}
\value{
A tibble data frame.
}
\description{
Retrieve a list of available BIS data sets
}
\examples{
\donttest{
ds <- get_datasets()
}
}
